corporate_financial_data_pipeline/requirements.txt
langchain
defaults
langchain-community
langchain-nomic
langchain-pinecone
redis
rq
fastapi
streamlit
pinecone-client
langchain-groq
anyio
aioredis
prometheus_client 